backend/requirements.txt
fastapi
uvicorn
spacy
python-multipart
scikit-learn
sentence-transformers
pdfminer.six
pytest
